+++ Tests/SFEngineTests/SFEngineTests-Bridging-Header.h
@@ -1,0 +1,2 @@
+// SFEngineTests-Bridging-Header.h
+#import "SFEngine.h"